project/requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Machine Learning & Embeddings
torch>=2.0.0
sentence-transformers==2.2.2
transformers>=4.30.0

# Vector Store
faiss-cpu==1.7.4
# For GPU support, use: faiss-gpu==1.7.4

# LLM Integration
google-generativeai==0.3.2

# LiveKit (Voice Agent)
livekit==1.0.19
livekit-agents==1.3.2
livekit-api==1.0.7
livekit-plugins-google==1.3.2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.20

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Logging & Monitoring
loguru==0.7.2

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Optional: Advanced Features
# For reranking (if implemented)
# rank-bm25==0.2.2
# cross-encoder==0.1.0

# Optional: Data Validation
# pydantic-settings==2.0.3

# Development Tools (optional)
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.0
# pre-commit==3.5.0